refactor: [Android] - Auth > User 변경
